Moved `mock-api/` to `util/api/`
